web-app/backend/script/requirements.txt
ensemble_boxes==1.0.9
fastapi==0.95.1
imageio==2.28.1
ipython==8.13.2
keras==2.12.0
loguru==0.7.0
matplotlib==3.7.1
mmcv_full==1.7.1
mmdet==3.0.0
numpy==1.23.5
opencv_python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
Pillow==9.4.0
Pillow==9.5.0
pycocotools==2.0
pydantic==1.10.7
PyYAML==6.0
PyYAML==6.0
scipy==1.9.1
segmentation_models==1.0.1
six==1.16.0
skimage==0.0
tensorflow==2.12.0
tifffile==2023.4.12
torch==2.0.0
tqdm==4.65.0
